requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
uWSGI==2.0.18
python-magic==0.4.15
